requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
streamlit-extras==0.3.6

# AI and API Dependencies
google-generativeai==0.3.2
openai==1.3.7

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Caching
redis==5.0.1
cachetools==5.3.2

# HTTP and API
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development and Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Deployment
gunicorn==21.2.0

# Additional Streamlit Components
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-ace==0.1.1

# Utilities
python-dateutil==2.8.2
click==8.1.7
rich==13.7.0 